+++ types/shared/archiveThumbnails.d.ts
@@ -1,0 +1,9 @@
+/**
+ * Archive Thumbnails.
+ *
+ * @export
+ * @param {string} archiveDir Directory to archive.
+ * @return {string}           The archive path created.
+ */
+declare const _default: (archiveDir: string) => string;
+export default _default;